Backend/learn-dsa-backend/requirements.txt
fastapi
uvicorn
firebase-admin
cloudinary
python-dotenv